backend/requirements.txt
# FastAPI Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-multipart>=0.0.9

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Database
sqlalchemy>=2.0.36
alembic>=1.14.0
motor>=3.3.0          # MongoDB async driver
pymongo>=4.6.0        # MongoDB sync utilities

# Redis for caching/sessions (optional)
# redis==5.0.1
# hiredis==2.3.2

# Email
python-dotenv==1.0.0
# emails==0.6

# ML & Data Science (optional - comment out if not needed)
# numpy==1.26.3
# pandas==2.1.4
# scikit-learn==1.4.0
pdfplumber>=0.11.0
python-docx>=1.1.0
PyPDF2>=3.0.0

# HTTP Client
httpx>=0.27.0
aiofiles>=23.2.1

# Monitoring & Logging
python-json-logger==2.0.7

# Rate Limiting
slowapi==0.1.9

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
